refactor: move Island to shared components folder
