+++ src/components/IslandHeader.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from "react"
-import { cn } from "../../../../utils/cn"
+import { cn } from "../utils/cn"
 
 export function IslandHeader({
   heading,
